+++ mysite/urls.py
@@ -18,6 +18,6 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path("", include("polls.urls")),
+    path("polls/", include("polls.urls")),
     path('admin/', admin.site.urls),
 ]
