+++ polls/tests/__init__.py
@@ -1,1 +1,2 @@
 from .model import *
+from .url import *
